use zustand for saving user data
